DjangoProject/requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
gunicorn==20.1.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
sqlparse==0.4.4
tomli==2.0.1

pip==23.1.2
setuptools==67.8.0